+++ common/src/main/java/com/slymask3/instantblocks/reference/Names.java
@@ -30,5 +30,6 @@ public class Names {
         public static final String IB_WAND_GOLD = "wand_gold";
         public static final String IB_WAND_DIAMOND = "wand_diamond";
         public static final String IB_WAND_NETHERITE = "wand_netherite";
+        public static final String IB_WAND_CLEAR = "wand_clear";
     }
 }